requirements.txt
pyyaml
robotframework
robotframework-requests